move some thangs
